Rewrote unifiedDataLoader into typscript functions - Moved UnifiedDataLoader to a separate typescript file. - The method is no longer an angular service, and instead exports two 'getUnified' methods. - Updated other files that rely on this service. - Tested getUnifiedMessagesForoInterval, as used in enketoHelper.
