+++ www/js/diary/services.js
@@ -1,15 +1,15 @@
 'use strict';
 
 import angular from 'angular';
-import { getBaseModeByKey, getBaseModeOfLabeledTrip } from './diaryHelper';
 import { SurveyOptions } from '../survey/survey';
 import { getConfig } from '../config/dynamicConfig';
 import { getRawEntries } from '../commHelper';
+import { getUnifiedSensorDataForInterval, getUnifiedMessagesForInterval } from '../unifiedDataLoader'
 
 angular.module('emission.main.diary.services', ['emission.plugin.logger',
                                                 'emission.services'])
 .factory('Timeline', function($http, $ionicLoading, $ionicPlatform, $window,
-    $rootScope, UnifiedDataLoader, Logger, $injector) {
+    $rootScope, Logger, $injector) {
     var timeline = {};
     // corresponds to the old $scope.data. Contains all state for the current
     // day, including the indication of the current day
@@ -232,7 +232,7 @@ angular.module('emission.main.diary.services', ['emission.plugin.logger',
       Logger.log("About to pull location data for range "
         + moment.unix(tripStartTransition.data.ts).toString() + " -> " 
         + moment.unix(tripEndTransition.data.ts).toString());
-      return UnifiedDataLoader.getUnifiedSensorDataForInterval("background/filtered_location", tq).then(function(locationList) {
+      return getUnifiedSensorDataForInterval("background/filtered_location", tq).then(function(locationList) {
           if (locationList.length == 0) {
             return undefined;
           }
@@ -304,7 +304,7 @@ angular.module('emission.main.diary.services', ['emission.plugin.logger',
        }
        Logger.log("about to query for unprocessed trips from "
          +moment.unix(tq.startTs).toString()+" -> "+moment.unix(tq.endTs).toString());
-       return UnifiedDataLoader.getUnifiedMessagesForInterval("statemachine/transition", tq)
+       return getUnifiedMessagesForInterval("statemachine/transition", tq)
         .then(function(transitionList) {
           if (transitionList.length == 0) {
             Logger.log("No unprocessed trips. yay!");
