+++ www/js/services.js
@@ -4,87 +4,6 @@ import angular from 'angular';
 import { getRawEntries } from './commHelper';
 
 angular.module('emission.services', ['emission.plugin.logger'])
-
-.service('UnifiedDataLoader', function($window, Logger) {
-    var combineWithDedup = function(list1, list2) {
-      var combinedList = list1.concat(list2);
-      return combinedList.filter(function(value, i, array) {
-        var firstIndexOfValue = array.findIndex(function(element, index, array) {
-          return element.metadata.write_ts == value.metadata.write_ts;
-        });
-        return firstIndexOfValue == i;
-      });
-    };
-
-    // TODO: generalize to iterable of promises
-    var combinedPromise = function(localPromise, remotePromise, combiner) {
-        return new Promise(function(resolve, reject) {
-          var localResult = [];
-          var localError = null;
-
-          var remoteResult = [];
-          var remoteError = null;
-
-          var localPromiseDone = false;
-          var remotePromiseDone = false;
-
-          var checkAndResolve = function() {
-            if (localPromiseDone && remotePromiseDone) {
-              // time to return from this promise
-              if (localError && remoteError) {
-                reject([localError, remoteError]);
-              } else {
-                Logger.log("About to dedup localResult = "+localResult.length
-                    +"remoteResult = "+remoteResult.length);
-                var dedupedList = combiner(localResult, remoteResult);
-                Logger.log("Deduped list = "+dedupedList.length);
-                resolve(dedupedList);
-              }
-            }
-          };
-
-          localPromise.then(function(currentLocalResult) {
-            localResult = currentLocalResult;
-            localPromiseDone = true;
-          }, function(error) {
-            localResult = [];
-            localError = error;
-            localPromiseDone = true;
-          }).then(checkAndResolve);
-
-          remotePromise.then(function(currentRemoteResult) {
-            remoteResult = currentRemoteResult;
-            remotePromiseDone = true;
-          }, function(error) {
-            remoteResult = [];
-            remoteError = error;
-            remotePromiseDone = true;
-          }).then(checkAndResolve);
-        })
-    }
-
-    // TODO: Generalize this to work for both sensor data and messages
-    // Do we even need to separate the two kinds of data?
-    // Alternatively, we can maintain another mapping between key -> type
-    // Probably in www/json...
-    this.getUnifiedSensorDataForInterval = function(key, tq) {
-        var localPromise = $window.cordova.plugins.BEMUserCache.getSensorDataForInterval(key, tq, true);
-        var remotePromise = getRawEntries([key], tq.startTs, tq.endTs)
-          .then(function(serverResponse) {
-            return serverResponse.phone_data;
-          });
-        return combinedPromise(localPromise, remotePromise, combineWithDedup);
-    };
-
-    this.getUnifiedMessagesForInterval = function(key, tq, withMetadata) {
-      var localPromise = $window.cordova.plugins.BEMUserCache.getMessagesForInterval(key, tq, true);
-      var remotePromise = getRawEntries([key], tq.startTs, tq.endTs)
-          .then(function(serverResponse) {
-            return serverResponse.phone_data;
-          });
-      return combinedPromise(localPromise, remotePromise, combineWithDedup);
-    }
-})
 .service('ControlHelper', function($window,
                                    $ionicPopup,
                                    Logger) {
@@ -198,4 +117,4 @@ angular.module('emission.services', ['emission.plugin.logger'])
       return window.cordova.plugins.BEMConnectionSettings.getSettings();
     };
 
-})
+});
